refactor: move env.ts to configs/env.ts
